simpify the resume config and the trainer code
