+++ main/train.py
@@ -33,18 +33,13 @@ logger = logging.getLogger('global')
 
 def make_parser():
     parser = argparse.ArgumentParser(description='Test')
-    parser.add_argument('--config',
+    parser.add_argument('-cfg','--config',
                         default='',
                         type=str,
                         help='path to experiment configuration')
-    parser.add_argument('--resume-from-epoch',
+    parser.add_argument('-r', '--resume',
                         default=-1,
-                        type=int,
-                        help=r"latest completed epoch's number (from which training resumes)")
-    parser.add_argument('--resume-from-file',
-                        default="",
-                        type=str,
-                        help=r"latest completed epoch's snapshot file (from which training resumes)")
+                        help=r"completed epoch's number, latest or one model path")
 
     return parser
 
@@ -79,7 +74,7 @@ if __name__ == '__main__':
     # build trainer
     trainer = engine_builder.build(task, task_cfg.trainer, "trainer", optimizer,
                                    dataloader)
-    trainer.resume(parsed_args.resume_from_epoch, parsed_args.resume_from_file)
+    trainer.resume(parsed_args.resume)
     # trainer.init_train()
     logger.info("Start training")
     while not trainer.is_completed():
