+++ videoanalyst/engine/tester/tester_impl/vot.py
@@ -44,10 +44,9 @@ class VOTTester(TesterBase):
 
     extra_hyper_params = dict(
         device_num=1,
-        data_root=CfgNode(dict(
-            VOT2018="datasets/VOT/vot2018",
-            VOT2019="datasets/VOT/vot2019"
-        )),
+        data_root=CfgNode(
+            dict(VOT2018="datasets/VOT/vot2018",
+                 VOT2019="datasets/VOT/vot2019")),
         dataset_names=[
             "VOT2018",
         ],
